Use FipFrame as the base data record for fip data
